Updated to allow better support for UNQ type inheritance.  Support should now be present for using a derived UNQ type when a base type is expected. (this hasn't been tested since none of the code expects this at the moment).
Added support for importing another DTD into this file using an import tag.
Syntax is ~import{filename} at the top level.  There are now two top level
tags possible for DTDs -- ~dtd and ~import.

Also added support for multiple DTDs per file (these are concatenated together,
and stored in one namespace.  Not entirely sure what this is good for, but
it fell out of the support for the import tag.